# Fix fields merging and improve performance