--- a/core/src/main/scala/caliban/execution/Field.scala
+++ b/core/src/main/scala/caliban/execution/Field.scala
@@ -120,7 +120,7 @@ object Field {
           }
         case _                                                                                                      =>
       }
-      Field("", fieldType, None, fields = fieldList.result().toList)
+      Field("", fieldType, None, fields = fieldList.toList)
     }
 
     loop(selectionSet, fieldType).copy(directives = directives)