Fields were not properly merged during execution. I moved the merging to the validation phase and added some performance optimizations.

Before:
```
[info] Benchmark                               Mode  Cnt      Score      Error  Units
[info] GraphQLBenchmarks.fragmentsCaliban     thrpt    5     27.739 ±    2.741  ops/s
[info] GraphQLBenchmarks.fragmentsSangriaNew  thrpt    5     24.269 ±    4.216  ops/s
[info] GraphQLBenchmarks.fragmentsSangriaOld  thrpt    5      0.165 ±    0.011  ops/s
[info] GraphQLBenchmarks.introspectCaliban    thrpt    5   1585.284 ±  183.560  ops/s
[info] GraphQLBenchmarks.introspectSangria    thrpt    5    499.669 ±   19.512  ops/s
[info] GraphQLBenchmarks.simpleCaliban        thrpt    5  24539.809 ± 1188.386  ops/s
[info] GraphQLBenchmarks.simpleSangria        thrpt    5   7551.214 ± 1423.378  ops/s
```
After:
```
[info] Benchmark                               Mode  Cnt      Score     Error  Units
[info] GraphQLBenchmarks.fragmentsCaliban     thrpt    5     44.638 ±   1.523  ops/s
[info] GraphQLBenchmarks.fragmentsSangriaNew  thrpt    5     28.994 ±  16.953  ops/s
[info] GraphQLBenchmarks.fragmentsSangriaOld  thrpt    5      0.155 ±   0.039  ops/s
[info] GraphQLBenchmarks.introspectCaliban    thrpt    5   1782.754 ± 309.681  ops/s
[info] GraphQLBenchmarks.introspectSangria    thrpt    5    500.839 ±  15.303  ops/s
[info] GraphQLBenchmarks.simpleCaliban        thrpt    5  26091.734 ± 305.640  ops/s
[info] GraphQLBenchmarks.simpleSangria        thrpt    5   7831.910 ± 139.665  ops/s
```